Update README to be simpler and less intimidating.
Just based on the theory that, if we throw a book at them on how to prepare to begin, we can't expect anyone to even crack the cover. Keep it short and resourceful for later use, but they don't need to pour in hours of prep work here. 

Would gladly discuss changes here, as I did touch on a good bit of formatting.